Update static file copying to match current JSdoc default template implementation.
Makes static file copying match the documented behaviour.

And more importantly, fixes bug where if the directory separator used in
the config didn't match the directory separator of the OS it would attempt
to copy the static files over themselves, and due to the custom copyFileSync
implementation in jsdoc they would get clobbered to a 0 byte file.